move endpoint to coauthors file
