+++ openlibrary/asgi_app.py
@@ -99,103 +99,18 @@ def create_app() -> FastAPI:
     def health() -> dict[str, str]:
         return {"status": "ok"}
 
-    @app.get("/_fast/author/{olid}")
-    def author_coauthors(olid: str):
-        """Return author basic info and co-authors via Solr aggregation.
-
-        This is intentionally JSON-first for an initial endpoint. We can add an
-        HTML template later.
-        """
-        try:
-            data = _compute_author_network(olid)
-            return JSONResponse(data)
-        except HTTPException:
-            raise
-        except Exception as e:
-            logger.exception("error in author network endpoint for %s", olid)
-            raise HTTPException(status_code=500, detail=str(e))
-
     # Mount authors router
     from openlibrary.fastapi.authors import router as authors_router  # type: ignore
+    from openlibrary.fastapi.coauthors import router as coauthors_router  # type: ignore
 
     app.include_router(authors_router)
+    app.include_router(coauthors_router)
 
     # Finally, mount the legacy app at "/" so all existing routes keep working.
     # Ordering matters: Fast routes are declared BEFORE the catch-all mount.
     app.mount("/", WSGIMiddleware(legacy_wsgi))
 
     return app
-
-
-def _compute_author_network(olid: str) -> dict:
-    """Compute co-authors by querying Solr and author details via JSON API.
-
-    - Fetch author details from the classic JSON API: /authors/{olid}.json
-    - Query Solr for works with author_key:OL... and aggregate other authors.
-    """
-    import requests  # type: ignore
-
-    # Resolve service URLs from env with sensible defaults for docker-compose
-    base_web = os.environ.get("OL_WEB_INTERNAL", "http://localhost:8080")
-    solr_url = os.environ.get("SOLR_URL", "http://solr:8983/solr/openlibrary/select")
-
-    # Query Solr for works by this author and compute co-authors
-    params = {
-        "q": f"author_key:{olid}",
-        "fl": "key,title,author_key,author_name",
-        "rows": 2000,
-        "wt": "json",
-    }
-    s_resp = requests.get(solr_url, params=params, timeout=15)
-    s_resp.raise_for_status()
-    docs = s_resp.json().get("response", {}).get("docs", [])
-
-    if not docs:
-        raise HTTPException(
-            status_code=404, detail="author not found or no works in index"
-        )
-
-    # Derive canonical author name from docs where this olid appears
-    name_counts: dict[str, int] = {}
-    for d in docs:
-        keys: list[str] = d.get("author_key", []) or []
-        names: list[str] = d.get("author_name", []) or []
-        for i, k in enumerate(keys):
-            if k == olid:
-                nm = names[i] if i < len(names) else None
-                if nm:
-                    name_counts[nm] = name_counts.get(nm, 0) + 1
-
-    author_name = max(name_counts, key=name_counts.get) if name_counts else None
-
-    # Aggregate co-authors
-    counts: dict[str, dict[str, object]] = {}
-    for d in docs:
-        keys: list[str] = d.get("author_key", []) or []
-        names: list[str] = d.get("author_name", []) or []
-        # zip may drop if lengths differ; use index lookup
-        name_by_key = {
-            k: names[i] if i < len(names) else None for i, k in enumerate(keys)
-        }
-        for k in keys:
-            if k == olid:
-                continue
-            if k not in counts:
-                counts[k] = {
-                    "key": f"/authors/{k}",
-                    "olid": k,
-                    "name": name_by_key.get(k),
-                    "shared_works": 0,
-                }
-            counts[k]["shared_works"] = int(counts[k]["shared_works"]) + 1
-
-    coauthors = sorted(counts.values(), key=lambda x: -int(x["shared_works"]))
-
-    return {
-        "author": {"key": f"/authors/{olid}", "olid": olid, "name": author_name},
-        "co_authors": coauthors,
-        "stats": {"works_considered": len(docs), "unique_coauthors": len(coauthors)},
-    }
 
 
 # The ASGI app instance Gunicorn/Uvicorn will serve
